% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EffectStars2-package.R
\docType{package}
\name{EffectStars2-package}
\alias{EffectStars2}
\alias{EffectStars2-package}
\title{Plot EffectStars}
\description{
The package provides functions for the method of EffectStars as proposed by
Tutz and Schauberger (2013). Beside the main function
\code{\link{effectstars}} there exist methods for special objects, for
example for \code{\link[VGAM]{vglm}}-objects.
}
\author{
Gunther Schauberger\cr
\url{www.stat.uni-muenchen.de\~schauberger}\cr
\email{gunther@stat.uni-muenchen.de}
}
\references{
Tutz, G. and Schauberger, G. (2013): \emph{Visualization of
Categorical Response Models - from Data Glyphs to Parameter Glyphs}, Journal
of Computational and Graphical Statistics, 22(1), 156--177
}
\seealso{
\code{\link{effectstars}}, \code{\link{star.ctrl}}, 
\code{\link{effectstars.vglm}}, \code{\link{effectstars.DIFlasso}}, 
\code{\link{effectstars.DIFboost}}
}
\keyword{package}

